+++ src/app/(list)/[teamId]/task-lists/[listId]/@side/(.)add-task/page.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const AddTask = () => <div>AddTask</div>;
+
+export default AddTask;
