--- src/app/(list)/[teamId]/task-lists/error.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-export default function Error() {
-  return <h1>에러 컴포넌트 만들어서 추가해요!</h1>;
-}
